Convert Rectangle to a Struct
I guess my main reservation about structs is that it enforces a distinction
between 'real' data in the struct versus virtual data accessed via functions
like Rectangle.area. Maybe this is just an OO bigotry that I need to discard.

I feel similarly about how CoffeeScript methods must be called with closing ().
I appreciate how CS allows simply getting a ref to the method when the () are
omitted, but I wonder if a similar language construct to Ruby or Elixir's & for
refs (and allowing the methods to be called without () as if they were a data
field) would be more to my liking in practice. I think so.

It would hide implementation details of the CS object from external users,
which seems like a clear win to me. How this relates to Elixir may be another
matter.